pyproject.toml
[project]
name = "pixelmemory-gateway"
version = "0.3.0"
description = "Give Claude a persistent, local-first memory. MCP server connecting Claude.ai to a local Pixeltable database."
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.8.0",
    "pixeltable>=0.2.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
pixelmemory-gateway = "src.server:main"
